Backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
jinja2
aiofiles
python-multipart
google-generativeai
PyJWT